Inline connectivity display helpers
