+++ SprinklerMobile/Stores/ConnectivityStore.swift
@@ -93,3 +93,52 @@ enum ConnectivityState: Equatable {
     case connected
     case offline(errorDescription: String?)
 }
+
+#if canImport(SwiftUI)
+
+/// User interface helpers that translate connectivity state into human readable
+/// copy and iconography. Keeping the values alongside the enum guarantees they
+/// are compiled into every platform target, preventing "no member" build
+/// errors when the Settings view references them.
+extension ConnectivityState {
+    /// Primary text describing the current connectivity state.
+    var statusTitle: String {
+        switch self {
+        case .connected:
+            return "Connected"
+        case .offline:
+            return "Offline"
+        }
+    }
+
+    /// Optional supporting text that surfaces troubleshooting information.
+    var statusMessage: String? {
+        switch self {
+        case .connected:
+            return "The controller is reachable on your network."
+        case let .offline(description):
+            return description ?? "Tap Run Health Check to troubleshoot the connection."
+        }
+    }
+
+    /// SF Symbol name used to visually represent the connectivity state.
+    var statusIcon: String {
+        switch self {
+        case .connected:
+            return "checkmark.circle.fill"
+        case .offline:
+            return "xmark.circle.fill"
+        }
+    }
+
+    /// Accent color that matches the status symbol and hero card chip.
+    var statusColor: Color {
+        switch self {
+        case .connected:
+            return .appSuccess
+        case .offline:
+            return .appDanger
+        }
+    }
+}
+#endif
